changed float to double precision; fixed testing errors introduced by adding ambigous residues
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@5434 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/CHEMISTRY/AASequence.C
@@ -906,8 +906,8 @@ namespace OpenMS
 					if (res_ptr == 0)
 					{
 						Residue res(tag, String(""), String(""), EmpiricalFormula(""));
-						res.setMonoWeight(tag.toFloat(), Residue::Internal);
-						res.setAverageWeight(tag.toFloat(), Residue::Internal);
+						res.setMonoWeight(tag.toDouble(), Residue::Internal);
+						res.setAverageWeight(tag.toDouble(), Residue::Internal);
 						getResidueDB_()->addResidue(res);
 						sequence.push_back(getResidueDB_()->getResidue(tag));
 					}
